requirements.txt
Django==2.2.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
Faker==2.0.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
